Added validation user and password

+++ app/routes/users.py
@@ -4,9 +4,9 @@ from typing import List
 from fastapi import APIRouter, Depends, Query, HTTPException
 from starlette import status
 
-from app.conf.messages import ERROR_ACCOUNT_EXISTS
+from app.conf.messages import ERROR_ACCESS
 from app.repository.dependencies import users_service
-from app.schemas.user_schemas import UserUpdate, UserBase, SignUpRequestModel
+from app.schemas.user_schemas import UserUpdate, UserBase, SignUpRequestModel, UserDetail
 from app.services.auth import auth_service
 from app.services.users import UsersService
 
@@ -21,7 +21,7 @@ async def SingUp(body: SignUpRequestModel, users_service: UsersService = Depends
     return f"User id:{user}"
 
 
-@route.get("/", response_model=List[UserBase])
+@route.get("/", response_model=List[UserDetail])
 async def get_users(
         limit: int = Query(10, le=300),
         offset: int = 0,
@@ -30,7 +30,7 @@ async def get_users(
     return await users_service.get_users(limit, offset)
 
 
-@route.get("/{user_id}", response_model=UserBase)
+@route.get("/{user_id}", response_model=UserDetail)
 async def read_user(user_id: int, users_service: UsersService = Depends(users_service)):
     return await users_service.get_user_by_id(user_id)
 
@@ -40,7 +40,11 @@ async def update_user(
         user_id: int,
         user_update: UserUpdate,
         users_service: UsersService = Depends(users_service),
+        current_user: dict = Depends(auth_service.get_current_user)
 ):
+    if current_user.id != user_id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=ERROR_ACCESS)
+    user_update.hashed_password = auth_service.get_password_hash(user_update.hashed_password)
     await users_service.update_user(user_id, user_update)
     logging.info(f"User {user_id} was changed")
     return await users_service.get_user_by_id(user_id)
@@ -48,8 +52,11 @@ async def update_user(
 
 @route.delete("/{user_id}", status_code=status.HTTP_204_NO_CONTENT)
 async def delete_user(
-        user_id: int, users_service: UsersService = Depends(users_service)
+        user_id: int, users_service: UsersService = Depends(users_service),
+        current_user: dict = Depends(auth_service.get_current_user)
 ):
+    if current_user.id != user_id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=ERROR_ACCESS)
     await users_service.delete_user(user_id)
     logging.info(f"User {user_id} was deleted")
     return f"User {user_id} was deleted"
